mail server changed to virtual specific server controlled by load balancer of support (ODD division) provided by F5.
Earlier it was (squmail) directed to exchange server load balancer. But due to intelligence issue, even if one of the exchange server not function properly, chances are there to push the mail to inactive server because of only round robin scheduling.

where as in present scenario, f-5 load balancer is having a pool of the listed exchange servers which keep noticing whether the server is responding or not and send the packets to only active one/s.